remove sorting in sidebar
